src/edp/app/pyproject.toml
# Copyright (C) 2025 Intel Corporation
# SPDX-License-Identifier: Apache-2.0

[project]
name = "Intel_Enterprise_RAG-EDP_Microservice"
version = "1.0.1"
requires-python = ">=3.10"

dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.30.1",
    "python-dotenv==1.0.1",
    "SQLAlchemy==2.0.36",
    "minio==7.2.12",
    "psycopg2-binary==2.9.10",
    "PyJWT==2.10.1",
    "celery==5.5.3",
    "redis==5.2.1",
    "flower==2.0.1",
    "validators==0.34.0",
    "requests==2.32.4",
    "prometheus-client==0.21.1",
    "boto3==1.36.26",
    "starlette==0.40.0",
    "jinja2==3.1.6",
    "pyyaml==6.0.2",
    "h11==0.16.0",
    "tornado==6.5.1",
    "pyjwt==2.10.1"
]

[project.optional-dependencies]
test = [
    "aiohttp==3.11.12",
    "shortuuid==1.0.13",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "fastapi==0.115.12",
    "docarray[full]==0.40.0",
    "opentelemetry-api==1.25.0",
    "opentelemetry-exporter-otlp==1.25.0",
    "opentelemetry-sdk==1.25.0",
]
